# Turn `memSeries.mmMaxTime` into a method