As the comment on this method states: we only use it during WAL replay. I was going to initially just build a temporary map in `processWALSamples` that would hold `mmMaxTime` valuesm but then I realized that we can just read it from `memSeries.mmappedChunks` as we assign the memory mmapped chunks there.

Maybe there's something I'm missing?