@bboreham 2024-07-15
The history is it went faster when I cached the value in #9160.