@colega 2024-07-15
I see, but aren't we paying an excessively high price for just going faster during WAL replay?

Anyway, let me run some `LoadWAL` benchmarks, also with my other proposal, and then we can talk.