@colega 2024-07-15
As you pointed out in Slack, this wouldn't work so nice if we [turned mmapedchunks](https://github.com/prometheus/prometheus/pull/12671) to a linked list. In that case [caching the value in a map could work better](https://github.com/prometheus/prometheus/pull/14473)

You're right about cache locality, but also benchmarks don't look _terrible_ (what is considered terrible here?) considering we're removing 8 bytes from `memSeries` struct.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                                                                                                                                                             │     main      │               method                │              mapcache               │
                                                                                                                                                             │    sec/op     │    sec/op     vs base               │    sec/op     vs base               │
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12          165.6m ±  4%   164.7m ±  1%       ~ (p=0.052 n=10)   166.6m ±  1%       ~ (p=0.853 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         168.0m ±  1%   167.6m ±  1%       ~ (p=0.247 n=10)   167.5m ±  2%       ~ (p=0.353 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         171.1m ±  0%   170.5m ±  1%       ~ (p=0.165 n=10)   170.5m ±  1%  -0.31% (p=0.023 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        198.0m ±  0%   197.6m ±  0%       ~ (p=0.315 n=10)   197.4m ±  1%       ~ (p=1.000 n=10)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12          220.8m ± 24%   220.2m ±  4%       ~ (p=0.631 n=10)   220.2m ±  2%       ~ (p=0.684 n=10)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12          243.1m ± 20%   235.6m ±  3%  -3.08% (p=0.043 n=10)   235.5m ±  3%  -3.12% (p=0.035 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12          79.11m ±  4%   79.04m ±  1%       ~ (p=0.912 n=10)   77.42m ±  0%  -2.13% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12          80.61m ±  1%   81.05m ±  0%       ~ (p=0.123 n=10)   79.53m ±  1%  -1.34% (p=0.002 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12          83.52m ±  1%   83.61m ±  1%       ~ (p=0.739 n=10)   81.85m ±  1%  -1.99% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        100.54m ± 11%   99.69m ±  2%       ~ (p=0.190 n=10)   97.88m ±  1%  -2.65% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12      832.9m ±  4%   859.6m ± 11%  +3.21% (p=0.009 n=10)   860.4m ±  2%  +3.30% (p=0.015 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12      845.0m ±  2%   870.7m ±  2%  +3.05% (p=0.001 n=10)   892.2m ±  4%  +5.60% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12      886.3m ± 16%   888.8m ±  2%       ~ (p=0.853 n=10)   911.6m ± 15%       ~ (p=0.063 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12      1.084 ±  1%    1.118 ±  1%  +3.15% (p=0.000 n=10)    1.133 ±  1%  +4.54% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12         279.3m ±  6%   279.5m ±  3%       ~ (p=0.853 n=10)   278.7m ±  8%       ~ (p=0.971 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12         280.9m ±  1%   281.8m ±  1%       ~ (p=0.247 n=10)   281.4m ±  0%       ~ (p=0.315 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12         284.5m ±  1%   283.9m ±  1%       ~ (p=0.393 n=10)   284.5m ±  4%       ~ (p=0.912 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        302.3m ±  2%   301.3m ±  1%       ~ (p=0.143 n=10)   301.0m ±  3%       ~ (p=0.315 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12          251.1m ±  1%   251.9m ±  6%       ~ (p=0.481 n=10)   250.2m ±  1%       ~ (p=0.165 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12          253.3m ±  4%   252.5m ±  1%       ~ (p=0.739 n=10)   253.1m ±  2%       ~ (p=0.971 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12          252.3m ±  1%   252.9m ±  1%       ~ (p=0.529 n=10)   256.1m ±  2%       ~ (p=0.143 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         278.5m ±  4%   267.4m ±  4%       ~ (p=0.105 n=10)   266.6m ±  1%  -4.27% (p=0.035 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12      3.263 ±  2%    3.355 ±  6%  +2.82% (p=0.019 n=10)    3.297 ±  1%       ~ (p=0.280 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12      3.279 ±  1%    3.357 ±  3%  +2.40% (p=0.000 n=10)    3.286 ±  1%       ~ (p=0.123 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12      3.289 ±  2%    3.442 ±  2%  +4.66% (p=0.000 n=10)    3.342 ±  2%  +1.63% (p=0.019 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12     3.510 ±  2%    3.578 ±  4%  +1.94% (p=0.015 n=10)    3.520 ±  2%       ~ (p=0.739 n=10)
geomean                                                                                                                                                         372.0m         373.7m        +0.47%                  372.5m        +0.13%
```