@colega 2024-07-29
Benchmarks on a 5GB WAL, just for the record:

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
               │    main    │               method               │
               │   sec/op   │   sec/op     vs base               │
LoadRealWLs-12   99.82 ± 6%   123.07 ± 9%  +23.29% (p=0.002 n=6)

               │     main     │               method               │
               │     B/op     │     B/op      vs base              │
LoadRealWLs-12   34.99Gi ± 0%   35.19Gi ± 0%  +0.58% (p=0.002 n=6)

               │    main     │              method               │
               │  allocs/op  │  allocs/op   vs base              │
LoadRealWLs-12   382.8M ± 0%   382.9M ± 0%  +0.03% (p=0.002 n=6)
```